Did some coding preparations for Ty.
Mason here. No this isn't my branch, this is Ty's branch. But he's not here now so I'm taking the wheel >:). Just kidding, but anyway I repurposed the testlevel movement code to be used for the player character if some revisions are done to it, but made sure to document how to reverse the changes and continue his work as it was. Also I changed the base node of the player to CharacterBody2D and added some of the possibly necessary things to get that working if he were to use it since it didn't seem like it was touched yesterday.