Added and fixed elements
Finally added proper buttons. They press the right way and can be disabled which turns them gray. The buttons are coded to go into the main and settings scenes. Quit button works. None of the buttons have words on them. Also added a camera so the buttons are much closer to the screen.